fix: update to deploy scripts
