+++ scripts/deploy/bond-deploy-no-etherscan.ts
@@ -1,0 +1,7 @@
+import {BondFactory, BondMediator} from '../../typechain-types'
+import {deployContractWithProxy} from '../../test/framework/contracts'
+
+export async function deployPerformanceBonds(): Promise<void> {
+    const factory = await deployContractWithProxy<BondFactory>('BondFactory')
+    await deployContractWithProxy<BondMediator>('BondMediator', factory.address)
+}
