+++ scripts/verify/whitelist-collateral.ts
